+++ bootique/src/main/java/io/bootique/annotation/Args.java
@@ -33,7 +33,6 @@ import java.lang.annotation.Target;
 @Target({ ElementType.PARAMETER, ElementType.FIELD, ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)
 @Qualifier
-@javax.inject.Qualifier
 public @interface Args {
 
 }