+++ bootique/src/main/java/io/bootique/annotation/DIConfigs.java
@@ -32,6 +32,5 @@ import java.lang.annotation.Target;
 @Target({ElementType.PARAMETER, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
 @Qualifier
-@javax.inject.Qualifier
 public @interface DIConfigs {
 }